requirements.txt
Django==1.7
django-imagekit==3.2.4
pilkit==1.1.12
Pillow==2.6.1
django-jsonview
gunicorn
whitenoise
django-debug-toolbar
django-pickling
coverage==3.7.1